+++ bw_processing/__init__.py
@@ -17,3 +17,4 @@ from .processed_package import (
     format_processed_resource,
 )
 from .loading import load_package
+from .filesystem import safe_filename, clean_datapackage_name