+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Changelog
 
-## [0.2.DEV]
+## [0.3] - 2021-05-19
 
 Nearly complete reconceptualization of the package structure and logic, based around the use of [PyFilesystem2](https://docs.pyfilesystem.org/en/latest/). Much more complete package, with better testing and documentation.